Implement reviewer comments - review 2702403
Reviewer Comments:
------------------
- Conclusions Phase
Ruberic: The results of the analysis are presented such that any
limitations are clear. The analysis does not state or imply that
one change causes another based solely on a correlation.

Reviewer comment : Please note that the report is expected to
include a discussion about the limitations and shortcomings of the
analysis and the dataset. For example, how many samples the data
include? Do you think that the number of samples is a good
representative of the population? Do you
 think that missing values might influence the results of the analysis?
The analysis includes comparisons between different categories, you
can discuss the importance of statistical test to appreciate the
significance of the differences. Optionally you can even implement
the appropriate statistical test.

- Communication
Ruberic: Visualizations made in the project depict the data in an
appropriate manner that allows plots to be readily interpreted.

Reviewer comment: Please add a discussion under each chart to
describe what each chart depicts. Please add the x and the y-axis
labels to all the charts. Please add a descriptive title for each
chart.

Project Changes:
---------------------
- Create detailed table of contents.
- Do minor fixes on script.
- Improve the plots.

> [Review 2702403 link](https://review.udacity.com/#!/reviews/2702403)